requirements.txt
# Core dependencies untuk Wave Detection System
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0

# Visualization
plotly>=5.17.0

# PDF Generation
reportlab>=4.0.0

# Twilio untuk WhatsApp & SMS
twilio>=8.10.0

# Environment variables
python-dotenv>=1.0.0

# HTTP Requests untuk BMKG API
requests>=2.31.0

